fix de varias cosas al mudar la l{ogica
