+++ recruitradar/appUsers/urls.py
@@ -4,7 +4,6 @@ from django.contrib.auth.views import LogoutView
 from django.urls import reverse_lazy
 #url
 urlpatterns = [
-    path('',views.home,name="Home"),
     path('login',views.login_request,name="Login"),
     path('logout',views.logout_request,name="Logout"),
     path('register',views.register,name="Register"),
