+++ recruitradar/recruitradar/settings.py
@@ -39,8 +39,9 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'appUsers',
     'appCurriculos',
+    'appUsers',
+    
 ]
 
 MIDDLEWARE = [
