+++ recruitradar/recruitradar/urls.py
@@ -21,8 +21,9 @@ from django.urls import path,include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('appUsers/',include('appUsers.urls')),
     path('appCurriculos/',include('appCurriculos.urls')),
+    path('appUsers/',include('appUsers.urls')),
+    
 ]
 
 urlpatterns += static(settings.MEDIA_URL,document_root=settings.MEDIA_ROOT)
